+++ packages/rum/src/boot/postStartStrategy.ts
@@ -4,7 +4,8 @@ import type { Telemetry, DeflateEncoder, SessionManager } from '@datadog/browser
 import { asyncRunOnReadyState, monitorError, Observable } from '@datadog/browser-core'
 import { getSessionReplayLink } from '../domain/getSessionReplayLink'
 import { startRecorderInitTelemetry } from '../domain/startRecorderInitTelemetry'
-import type { startRecording } from './startRecording'
+import type { startRecording } from './datadogRecorder'
+import { lazyLoadRecorder } from './lazyLoadRecorder'
 
 export type StartRecording = typeof startRecording
 
@@ -41,7 +42,6 @@ export function createPostStartStrategy(
   lifeCycle: LifeCycle,
   sessionManager: SessionManager,
   viewHistory: ViewHistory,
-  loadRecorder: () => Promise<StartRecording | undefined>,
   getOrCreateDeflateEncoder: () => DeflateEncoder | undefined,
   telemetry: Telemetry
 ): Strategy {
@@ -68,7 +68,7 @@ export function createPostStartStrategy(
     observable.notify({ type: 'start', forced })
 
     const [startRecordingImpl] = await Promise.all([
-      notifyWhenSettled(observable, { type: 'recorder-settled' }, loadRecorder()),
+      notifyWhenSettled(observable, { type: 'recorder-settled' }, lazyLoadRecorder()),
       notifyWhenSettled(observable, { type: 'document-ready' }, asyncRunOnReadyState(configuration, 'interactive')),
     ])