+++ packages/rum/src/boot/recorderApi.ts
@@ -21,11 +21,10 @@ import {
   startDeflateWorker,
 } from '../domain/deflate'
 import { isBrowserSupported } from './isBrowserSupported'
-import type { StartRecording } from './postStartStrategy'
 import { createPostStartStrategy } from './postStartStrategy'
 import { createPreStartStrategy } from './preStartStrategy'
 
-export function makeRecorderApi(loadRecorder: () => Promise<StartRecording | undefined>): RecorderApi {
+export function makeRecorderApi(): RecorderApi {
   if ((canUseEventBridge() && !bridgeSupports(BridgeCapability.RECORDS)) || !isBrowserSupported()) {
     return {
       start: noop,
@@ -100,7 +99,6 @@ export function makeRecorderApi(loadRecorder: () => Promise<StartRecording | und
       lifeCycle,
       sessionManager,
       viewHistory,
-      loadRecorder,
       getOrCreateDeflateEncoder,
       telemetry
     )