+++ test/e2e/scenario/rum/actions.scenario.ts
@@ -753,3 +753,41 @@ test.describe('custom actions with startAction/stopAction', () => {
       expect(relatedFetch).toBeDefined()
     })
 })
+
+test.describe('action collection with composed path selector', () => {
+  createTest('should not return a composed_path_selector if flag is disabled')
+    .withRum({ trackUserInteractions: true })
+    .withBody(html`
+      <button>Click</button>
+      <button id="my-button" data-test-allowed="test-btn" data-random="secret" class="foo bar baz">Click me</button>
+    `)
+    .run(async ({ intakeRegistry, flushEvents, page }) => {
+      const button = page.locator('#my-button')
+      await button.click()
+      await flushEvents()
+
+      const actionEvents = intakeRegistry.rumActionEvents
+      expect(actionEvents).toHaveLength(1)
+      expect(actionEvents[0]._dd.action?.target?.composed_path_selector).toBeUndefined()
+    })
+  createTest('should return a composed_path_selector if flag is enabled')
+    .withRum({
+      trackUserInteractions: true,
+      enableExperimentalFeatures: ['composed_path_selector'],
+    })
+    .withBody(html`
+      <button>Click</button>
+      <button id="my-button" data-test-id="test-btn" data-random="secret" class="foo bar baz">Click me</button>
+    `)
+    .run(async ({ intakeRegistry, flushEvents, page }) => {
+      const button = page.locator('#my-button')
+      await button.click()
+      await flushEvents()
+
+      const actionEvents = intakeRegistry.rumActionEvents
+      expect(actionEvents).toHaveLength(1)
+      expect(actionEvents[0]._dd.action?.target?.composed_path_selector).toBe(
+        'BUTTON#my-button[data-test-id="test-btn"].bar.baz.foo:nth-child(2):nth-of-type(2);'
+      )
+    })
+})